requirements.txt
joblib==1.4.0
gdown==5.2.0
pandas==2.2.2
numpy==2.2.6
scikit-learn==1.4.2
lightgbm==4.6.0
imbalanced-learn==0.14.0
fastapi==0.119.0
mlflow==3.4.0
PyYAML==6.0.2
uvicorn==0.37.0